ft: new mob finished - zombie

+++ Assets/Scripts/Flyweight/EnemyStats.cs
@@ -6,7 +6,7 @@ public class EnemyStats : ActorStats
     [SerializeField] private EnemyStatValues enemyStats;
 
     public int Damage => enemyStats.Damage;
-    public int AttackRange => enemyStats.AttackRange;
+    public float AttackRange => enemyStats.AttackRange;
 
     public float RestAfterAttack => enemyStats.RestAfterAttack;
 
@@ -16,6 +16,6 @@ public class EnemyStats : ActorStats
 public struct EnemyStatValues
 {
     public int Damage;
-    public int AttackRange;
+    public float AttackRange;
     public float RestAfterAttack;
 }
